fix: move parent tag icon folder to top level
