+++ apps/journeys-admin/src/components/TemplateView/TemplateTags/TemplateTags.tsx
@@ -7,9 +7,9 @@ import { Swiper, SwiperSlide } from 'swiper/react'
 import { JourneyFields_tags as Tag } from '@core/journeys/ui/JourneyProvider/__generated__/JourneyFields'
 
 import { useTagsQuery } from '../../../libs/useTagsQuery'
+import { ParentTagIcon } from '../../ParentTagIcon'
 
 import { getSortedTags } from './getSortedTags'
-import { ParentTagIcon } from './ParentTagIcon'
 import { TagItem } from './TagItem'
 
 interface TemplateTagsProps {
